requirements.txt
numpy==1.25.1
tensorflow-macos=2.12.0